hams-stock-stt-worker/requirements.txt
fastapi
uvicorn
yt-dlp
openai-whisper
torch
